+++ src/SimpleResize.cpp
@@ -13,12 +13,12 @@
 
 #if defined(MVTOOLS_X86)
 void simpleResize_uint8_t_avx2(const SimpleResize *simple,
-                               uint8_t *dstp, int dst_stride,
-                               const uint8_t *srcp, int src_stride,
+                               uint8_t *dstp, ptrdiff_t dst_stride,
+                               const uint8_t *srcp, ptrdiff_t src_stride,
                                int horizontal_vectors);
 void simpleResize_int16_t_avx2(const SimpleResize *simple,
-                               int16_t *dstp, int dst_stride,
-                               const int16_t *srcp, int src_stride,
+                               int16_t *dstp, ptrdiff_t dst_stride,
+                               const int16_t *srcp, ptrdiff_t src_stride,
                                int horizontal_vectors);
 #endif
 
@@ -28,8 +28,8 @@ static void InitTables(int *offsets, int *weights, int out, int in) {
     float leftmost = 0.5f;       // + shift
     float rightmost = in - 0.5f; // + shift
 
-    int leftmost_idx = std::max<int>(leftmost, 0);
-    int rightmost_idx = std::min<int>(rightmost, in - 1);
+    int leftmost_idx = std::max<int>(static_cast<int>(leftmost), 0);
+    int rightmost_idx = std::min<int>(static_cast<int>(rightmost), in - 1);
 
     for (int i = 0; i < out; i++) {
         float position = (i + 0.5f) * (float)in / (float)out;
@@ -58,8 +58,8 @@ static void InitTables(int *offsets, int *weights, int out, int in) {
 // Thread-safe.
 template <typename PixelType>
 static void simpleResize(const SimpleResize *simple,
-                         PixelType *dstp, int dst_stride,
-                         const PixelType *srcp, int src_stride,
+                         PixelType *dstp, ptrdiff_t dst_stride,
+                         const PixelType *srcp, ptrdiff_t src_stride,
                          int horizontal_vectors) {
 
     // Apparently only 16 bit vectors need limiting.
